hestia/requirements.txt
fastapi>=0.115.0
pydantic>=2.10.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
requests>=2.32.0
discord.py>=2.3.0